% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_vignette.R
\name{insert_front_matter}
\alias{insert_front_matter}
\title{insert vignette front matter into template and use new file}
\usage{
insert_front_matter(input, output, the_title)
}
\arguments{
\item{input}{`character`. Name of template file.}

\item{output}{`character`. Name of new      file.}

\item{the_title}{`character`. Title of the new vignette.}
}
\value{
whatever `ymlthis::use_rmarkdown` does
}
\description{
insert vignette front matter into template and use new file
}
