% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space_csl.R
\name{space_csl}
\alias{space_csl}
\title{Force item spacing in chosen Citation Style Language file}
\usage{
space_csl(
  space = "2",
  cslfile = "https://www.zotero.org/styles/american-political-science-association",
  destfile = "apsa.csl",
  fallback = FALSE
)
}
\arguments{
\item{space}{\code{character}. line spacing between entries. Default \code{"2"}.}

\item{cslfile}{\code{character}. Location of the CSL file. Default \code{"https://www.zotero.org/styles/american-political-science-association"}.}

\item{destfile}{\code{character}. Filename to save local copy of csl file to. Default \code{"apsa.csl"}.}

\item{fallback}{\code{  logical}. Use default if \code{cslfile} not found? Default \code{FALSE} (stops with an error).}
}
\value{
the argument \code{destfile}
}
\description{
Force item spacing in chosen Citation Style Language file
}
