% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_memo.R
\name{use_memo}
\alias{use_memo}
\title{Create a document in `arsedown::pdf_memo` format}
\usage{
use_memo(
  name,
  title = name,
  what = system.file("rmarkdown", "templates", "pdf-memo", "skeleton", "skeleton.Rmd",
    package = "arsedown"),
  whence = here::here(),
  overwrite = FALSE
)
}
\arguments{
\item{name}{File name to use for new vignette. Should consist only of
numbers, letters, \verb{_} and \code{-}. Lower case is recommended. Can include the
\code{".Rmd"} or \code{".qmd"} file extension, which also dictates whether to place
an R Markdown or Quarto vignette. R Markdown (\code{".Rmd"}) is the current
default, but it is anticipated that Quarto (\code{".qmd"}) will become the
default in the future.}

\item{title}{The title of the vignette. If not provided, a title is generated
from \code{name}.}

\item{what}{character. Full path to a `.Rmd` skeleton. Defaults to the installed location of the `pdf_memo` skeleton.}

\item{whence}{`character`. A file path. Default `[here::here]()`.}

\item{overwrite}{logical. Overwrite an existing vignette with the same `name`? Default `FALSE` (gives an error).}
}
\value{
invisibly returns the full path to the new file
}
\description{
Creates a new R Markdown document in the current project directory using the `pdf_memo` skeleton.
}
