% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_vignette.R
\name{use_memo_vignette}
\alias{use_memo_vignette}
\title{Create a vignette in `arsedown::pdf_memo` format}
\usage{
use_memo_vignette(
  name,
  title = name,
  what = system.file("rmarkdown", "templates", "pdf-memo", "skeleton", "skeleton.Rmd",
    package = "arsedown"),
  overwrite = FALSE
)
}
\arguments{
\item{name}{Base for file name to use for new vignette. Should consist only
of numbers, letters, \verb{_} and \code{-}. Lower case is recommended.}

\item{title}{The title of the vignette.}

\item{what}{character. Full path to a `.Rmd` skeleton. Defaults to the installed location of the `pdf_memo` skeleton.}

\item{overwrite}{logical. Overwrite an existing vignette with the same `name`? Default `FALSE` (gives an error).}
}
\value{
invisibly returns the full path to the new file
}
\description{
Creates a new vignette in `vignettes/` using the `pdf_memo` skeleton.
}
