% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_memo.R
\name{pdf_memo}
\alias{pdf_memo}
\title{wrapper for \code{bookdown::\link[bookdown]{pdf_book}} with Gordon Arsenoff's defaults}
\usage{
pdf_memo(smolhead = TRUE, ...)
}
\arguments{
\item{smolhead}{\code{logical}. Smaller-sized headings?   Default \code{TRUE} (\code{FALSE} gives regular LaTeX headings).}

\item{...}{Other arguments to be passed to \code{base_format}.}
}
\value{
whatever \code{pdf_book} does
}
\description{
wrapper for \code{bookdown::\link[bookdown]{pdf_book}} with Gordon Arsenoff's defaults
}
